pre version 0.0.1

+++ core/db_model.go
@@ -3,7 +3,7 @@ package core
 type DbBaseModel struct {
 	Id          string `json:"id" gorm:"primaryKey;size:32"`
 	TenantId    string `json:"tenant_id" gorm:"size:32"`
-	DataVersion int    `json:"data_version"`
+	DataVersion int64  `json:"data_version"`
 	DataStatus  int    `json:"data_status"`
 	CreateTime  int64  `json:"create_time" gorm:"autoCreateTime:milli"`
 	UpdateTime  int64  `json:"update_time" gorm:"autoUpdateTime:milli"`
